- Feature search form language

+++ resources/views/admin/language/index.blade.php
@@ -24,6 +24,16 @@
         </h3>
     </div>
 
+    <div class='bound-search'>
+        {!! Form::open(array('method'=>'GET', 'route' => 'languages.index', 'class'=>'form navbar-form navbar-right searchform')) !!}
+        {!! Form::text('keywords', Request::get('keywords'),
+                                array(null,
+                                'class'=>'form-control',
+                                'placeholder'=>'Search for a tutorial...')) !!}
+            {!! Form::submit('Search', array('class'=>'btn btn-default')) !!}
+        {!! Form::close() !!}
+    </div>
+
     <table id="table" class="table table-striped table-hover">
         <thead>
         <tr>
@@ -41,7 +51,7 @@
                     <td><img class="flag flag-{{ $language->lang_code }}" alt="" /></td>
                     <td>
                       <a class='btn btn-primary' href="{!!  url('admin/languages/'.$language->id.'/edit') !!}">Edit</a> 
-                      {{ Form::open(['method' => 'DELETE', 'class'=>'inline', 'route' => ['languages.destroy', $language->id]]) }}
+                      {{ Form::open(['method' => 'DELETE', 'class'=>'inline', 'route' => ['languages.destroy', $language->id],'onsubmit' => 'return confirm("Are you sure ?")']) }}
                         {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
                       {{ Form::close() }}
                     </td>
@@ -49,6 +59,10 @@
             @endforeach
         </tbody>
     </table>
+
+    <div class='pagination'>
+        {!! $languages->appends(['keywords' => Request::get('keywords')])->render() !!}
+    </div>
 @endsection
 
 {{-- Scripts --}}
